+++ rust/operator-binary/src/controller.rs
@@ -1,9 +1,15 @@
 //! Ensures that `Pod`s are configured and running for each [`OpaCluster`]
 
 use crate::discovery::{self, build_discovery_configmaps};
+use crate::product_logging::{
+    extend_role_group_config_map, resolve_vector_aggregator_address, BundleBuilderLogLevel,
+    OpaLogLevel,
+};
 
 use snafu::{OptionExt, ResultExt, Snafu};
-use stackable_opa_crd::{OpaCluster, OpaConfig, OpaRole, APP_NAME, OPERATOR_NAME};
+use stackable_opa_crd::{Container, OpaCluster, OpaConfig, OpaRole, APP_NAME, OPERATOR_NAME};
+use stackable_operator::builder::VolumeBuilder;
+use stackable_operator::product_logging::spec::{AppenderConfig, LogLevel};
 use stackable_operator::{
     builder::{ConfigMapBuilder, ContainerBuilder, FieldPathEnvVar, ObjectMetaBuilder, PodBuilder},
     commons::product_image_selection::ResolvedProductImage,
@@ -11,13 +17,14 @@ use stackable_operator::{
         api::{
             apps::v1::{DaemonSet, DaemonSetSpec},
             core::v1::{
-                ConfigMap, ConfigMapVolumeSource, EmptyDirVolumeSource, EnvVar, HTTPGetAction,
-                PodSecurityContext, Probe, Service, ServiceAccount, ServicePort, ServiceSpec,
-                Volume,
+                ConfigMap, EnvVar, HTTPGetAction, PodSecurityContext, Probe, Service,
+                ServiceAccount, ServicePort, ServiceSpec,
             },
             rbac::v1::{ClusterRole, RoleBinding, RoleRef, Subject},
         },
-        apimachinery::pkg::{apis::meta::v1::LabelSelector, util::intstr::IntOrString},
+        apimachinery::pkg::{
+            api::resource::Quantity, apis::meta::v1::LabelSelector, util::intstr::IntOrString,
+        },
         Resource,
     },
     kube::runtime::{controller::Action, reflector::ObjectRef},
@@ -25,6 +32,10 @@ use stackable_operator::{
     logging::controller::ReconcilerError,
     product_config::{types::PropertyNameKind, ProductConfigManager},
     product_config_utils::{transform_all_roles_to_config, validate_all_roles_and_groups_config},
+    product_logging::{
+        self,
+        spec::{AutomaticContainerLogConfig, ContainerLogConfig, ContainerLogConfigChoice},
+    },
     role_utils::RoleGroupRef,
 };
 use std::{
@@ -46,7 +57,29 @@ pub const BUNDLES_INCOMING_DIR: &str = "/bundles/incoming";
 pub const BUNDLES_TMP_DIR: &str = "/bundles/tmp";
 pub const BUNDLE_BUILDER_PORT: i32 = 3030;
 
+const CONFIG_VOLUME_NAME: &str = "config";
+const CONFIG_DIR: &str = "/stackable/config";
+const LOG_VOLUME_NAME: &str = "log";
+const LOG_DIR: &str = "/stackable/log";
+const BUNDLES_VOLUME_NAME: &str = "bundles";
+const BUNDLES_DIR: &str = "/bundles";
+
 const DOCKER_IMAGE_BASE_NAME: &str = "opa";
+
+// ~ 5 MB
+const MAX_OPA_BUNDLE_BUILDER_LOG_FILE_SIZE_IN_BYTES: u32 = 5000000;
+const OPA_ROLLING_BUNDLE_BUILDER_LOG_FILES: u32 = 2;
+// ~ 5 MB
+const MAX_OPA_LOG_FILE_SIZE_IN_BYTES: u32 = 5000000;
+const OPA_ROLLING_LOG_FILES: u32 = 2;
+// ~ 1 MB
+const MAX_PREPARE_LOG_FILE_SIZE_IN_BYTES: u32 = 1000000;
+
+const LOG_FILE_VOLUME_SIZE_IN_MB: u32 = ((MAX_OPA_BUNDLE_BUILDER_LOG_FILE_SIZE_IN_BYTES
+    * OPA_ROLLING_BUNDLE_BUILDER_LOG_FILES)
+    + (MAX_OPA_LOG_FILE_SIZE_IN_BYTES * OPA_ROLLING_LOG_FILES)
+    + MAX_PREPARE_LOG_FILE_SIZE_IN_BYTES)
+    / 1000000;
 
 pub struct Ctx {
     pub client: stackable_operator::client::Client,
@@ -114,6 +147,19 @@ pub enum Error {
     },
     #[snafu(display("failed to resolve and merge config for role and role group"))]
     FailedToResolveConfig { source: stackable_opa_crd::Error },
+    #[snafu(display("illegal container name"))]
+    IllegalContainerName {
+        source: stackable_operator::error::Error,
+    },
+    #[snafu(display("failed to resolve the Vector aggregator address"))]
+    ResolveVectorAggregatorAddress {
+        source: crate::product_logging::Error,
+    },
+    #[snafu(display("failed to add the logging configuration to the ConfigMap [{cm_name}]"))]
+    InvalidLoggingConfig {
+        source: crate::product_logging::Error,
+        cm_name: String,
+    },
 }
 type Result<T, E = Error> = std::result::Result<T, E>;
 
@@ -190,6 +236,10 @@ pub async fn reconcile_opa(opa: Arc<OpaCluster>, ctx: Arc<Ctx>) -> Result<Action
         .await
         .context(ApplyRoleRoleBindingSnafu)?;
 
+    let vector_aggregator_address = resolve_vector_aggregator_address(&opa, &client)
+        .await
+        .context(ResolveVectorAggregatorAddressSnafu)?;
+
     for (rolegroup_name, rolegroup_config) in role_server_config.iter() {
         let rolegroup = RoleGroupRef {
             cluster: opa_ref.clone(),
@@ -201,8 +251,13 @@ pub async fn reconcile_opa(opa: Arc<OpaCluster>, ctx: Arc<Ctx>) -> Result<Action
             .merged_config(&OpaRole::Server, &rolegroup)
             .context(FailedToResolveConfigSnafu)?;
 
-        let rg_configmap =
-            build_server_rolegroup_config_map(&opa, &resolved_product_image, &rolegroup)?;
+        let rg_configmap = build_server_rolegroup_config_map(
+            &opa,
+            &resolved_product_image,
+            &rolegroup,
+            &merged_config,
+            vector_aggregator_address.as_deref(),
+        )?;
         let rg_daemonset = build_server_rolegroup_daemonset(
             &opa,
             &resolved_product_image,
@@ -382,8 +437,12 @@ fn build_server_rolegroup_config_map(
     opa: &OpaCluster,
     resolved_product_image: &ResolvedProductImage,
     rolegroup: &RoleGroupRef<OpaCluster>,
+    merged_config: &OpaConfig,
+    vector_aggregator_address: Option<&str>,
 ) -> Result<ConfigMap> {
-    ConfigMapBuilder::new()
+    let mut cm_builder = ConfigMapBuilder::new();
+
+    cm_builder
         .metadata(
             ObjectMetaBuilder::new()
                 .name_and_namespace(opa)
@@ -398,7 +457,19 @@ fn build_server_rolegroup_config_map(
                 ))
                 .build(),
         )
-        .add_data(CONFIG_FILE, build_config_file())
+        .add_data(CONFIG_FILE, build_config_file());
+
+    extend_role_group_config_map(
+        rolegroup,
+        vector_aggregator_address,
+        &merged_config.logging,
+        &mut cm_builder,
+    )
+    .context(InvalidLoggingConfigSnafu {
+        cm_name: rolegroup.object_name(),
+    })?;
+
+    cm_builder
         .build()
         .with_context(|_| BuildRoleGroupConfigSnafu {
             rolegroup: rolegroup.clone(),
@@ -435,13 +506,92 @@ fn build_server_rolegroup_daemonset(
             ..EnvVar::default()
         })
         .collect::<Vec<_>>();
-    let container_opa = ContainerBuilder::new("opa")
-        .expect("invalid hard-coded container name")
+
+    let prepare_container_name = Container::Prepare.to_string();
+    let mut cb_prepare =
+        ContainerBuilder::new(&prepare_container_name).context(IllegalContainerNameSnafu)?;
+
+    let bundle_builder_container_name = Container::BundleBuilder.to_string();
+    let mut cb_bundle_builder =
+        ContainerBuilder::new(&bundle_builder_container_name).context(IllegalContainerNameSnafu)?;
+
+    let opa_container_name = Container::Opa.to_string();
+    let mut cb_opa =
+        ContainerBuilder::new(&opa_container_name).context(IllegalContainerNameSnafu)?;
+
+    cb_prepare
         .image_from_product_image(resolved_product_image)
-        .command(build_opa_start_command())
+        .command(vec![
+            "bash".to_string(),
+            "-euo".to_string(),
+            "pipefail".to_string(),
+            "-c".to_string(),
+        ])
+        .args(vec![build_prepare_start_command(
+            merged_config,
+            &prepare_container_name,
+        )
+        .join(" && ")])
+        .add_volume_mount(BUNDLES_VOLUME_NAME, BUNDLES_DIR)
+        .add_volume_mount(LOG_VOLUME_NAME, LOG_DIR);
+
+    cb_bundle_builder
+        .image_from_product_image(resolved_product_image)
+        .command(vec![
+            "bash".to_string(),
+            "-euo".to_string(),
+            "pipefail".to_string(),
+            "-c".to_string(),
+        ])
+        .args(vec![build_bundle_builder_start_command(
+            merged_config,
+            &bundle_builder_container_name,
+        )])
+        .add_env_var_from_field_path("WATCH_NAMESPACE", FieldPathEnvVar::Namespace)
+        .add_env_var(
+            "OPA_BUNDLE_BUILDER_LOG",
+            bundle_builder_log_level(merged_config).to_string(),
+        )
+        .add_volume_mount(BUNDLES_VOLUME_NAME, BUNDLES_DIR)
+        .add_volume_mount(LOG_VOLUME_NAME, LOG_DIR)
+        .readiness_probe(Probe {
+            initial_delay_seconds: Some(5),
+            period_seconds: Some(10),
+            failure_threshold: Some(5),
+            http_get: Some(HTTPGetAction {
+                port: IntOrString::Int(BUNDLE_BUILDER_PORT),
+                path: Some("/status".to_string()),
+                ..HTTPGetAction::default()
+            }),
+            ..Probe::default()
+        })
+        .liveness_probe(Probe {
+            initial_delay_seconds: Some(30),
+            period_seconds: Some(10),
+            http_get: Some(HTTPGetAction {
+                port: IntOrString::Int(BUNDLE_BUILDER_PORT),
+                path: Some("/status".to_string()),
+                ..HTTPGetAction::default()
+            }),
+            ..Probe::default()
+        });
+
+    cb_opa
+        .image_from_product_image(resolved_product_image)
+        .command(vec![
+            "bash".to_string(),
+            "-euo".to_string(),
+            "pipefail".to_string(),
+            "-c".to_string(),
+        ])
+        .args(vec![build_opa_start_command(
+            merged_config,
+            &opa_container_name,
+        )])
         .add_env_vars(env)
         .add_container_port(APP_PORT_NAME, APP_PORT.into())
-        .add_volume_mount("config", "/stackable/config")
+        .add_volume_mount(CONFIG_VOLUME_NAME, CONFIG_DIR)
+        .add_volume_mount(LOG_VOLUME_NAME, LOG_DIR)
         .resources(merged_config.resources.to_owned().into())
         .readiness_probe(Probe {
             initial_delay_seconds: Some(5),
@@ -461,56 +611,56 @@ fn build_server_rolegroup_daemonset(
                 ..HTTPGetAction::default()
             }),
             ..Probe::default()
-        })
-        .build();
+        });
 
-    let container_bundle_builder = ContainerBuilder::new("opa-bundle-builder")
-        .expect("invalid hard-coded container name")
-        .image_from_product_image(resolved_product_image)
-        .command(vec![String::from("/stackable/opa-bundle-builder")])
-        .add_env_var_from_field_path("WATCH_NAMESPACE", FieldPathEnvVar::Namespace)
-        .add_volume_mount("bundles", "/bundles")
-        .readiness_probe(Probe {
-            initial_delay_seconds: Some(5),
-            period_seconds: Some(10),
-            failure_threshold: Some(5),
-            http_get: Some(HTTPGetAction {
-                port: IntOrString::Int(BUNDLE_BUILDER_PORT),
-                path: Some("/status".to_string()),
-                ..HTTPGetAction::default()
-            }),
-            ..Probe::default()
-        })
-        .liveness_probe(Probe {
-            initial_delay_seconds: Some(30),
-            period_seconds: Some(10),
-            http_get: Some(HTTPGetAction {
-                port: IntOrString::Int(BUNDLE_BUILDER_PORT),
-                path: Some("/status".to_string()),
-                ..HTTPGetAction::default()
-            }),
-            ..Probe::default()
-        })
-        .build();
+    let mut pb = PodBuilder::new();
 
-    let init_container = ContainerBuilder::new("init-container")
-        .expect("invalid hard-coded container name")
-        .image_from_product_image(resolved_product_image)
-        .command(vec!["bash".to_string()])
-        .args(vec![
-            "-euo".to_string(),
-            "pipefail".to_string(),
-            "-x".to_string(),
-            "-c".to_string(),
-            [
-                format!("mkdir -p {BUNDLES_ACTIVE_DIR}"),
-                format!("mkdir -p {BUNDLES_INCOMING_DIR}"),
-                format!("mkdir -p {BUNDLES_TMP_DIR}"),
-            ]
-            .join(" && "),
-        ])
-        .add_volume_mount("bundles", "/bundles")
-        .build();
+    pb.metadata_builder(|m| {
+        m.with_recommended_labels(build_recommended_labels(
+            opa,
+            &resolved_product_image.app_version_label,
+            &rolegroup_ref.role,
+            &rolegroup_ref.role_group,
+        ))
+    })
+    .add_init_container(cb_prepare.build())
+    .add_container(cb_opa.build())
+    .add_container(cb_bundle_builder.build())
+    .image_pull_secrets_from_product_image(resolved_product_image)
+    .add_volume(
+        VolumeBuilder::new(CONFIG_VOLUME_NAME)
+            .with_config_map(rolegroup_ref.object_name())
+            .build(),
+    )
+    .add_volume(
+        VolumeBuilder::new(BUNDLES_VOLUME_NAME)
+            .with_empty_dir(None::<String>, None)
+            .build(),
+    )
+    .add_volume(
+        VolumeBuilder::new(LOG_VOLUME_NAME)
+            .with_empty_dir(
+                None::<String>,
+                Some(Quantity(format!("{LOG_FILE_VOLUME_SIZE_IN_MB}Mi"))),
+            )
+            .build(),
+    )
+    .service_account_name(sa_name)
+    .security_context(PodSecurityContext {
+        run_as_user: Some(1000),
+        run_as_group: Some(1000),
+        fs_group: Some(1000),
+        ..PodSecurityContext::default()
+    });
+
+    if merged_config.logging.enable_vector_agent {
+        pb.add_container(product_logging::framework::vector_container(
+            resolved_product_image,
+            CONFIG_VOLUME_NAME,
+            LOG_VOLUME_NAME,
+            merged_config.logging.containers.get(&Container::Vector),
+        ));
+    }
 
     Ok(DaemonSet {
         metadata: ObjectMetaBuilder::new()
@@ -535,40 +685,7 @@ fn build_server_rolegroup_daemonset(
                 )),
                 ..LabelSelector::default()
             },
-            template: PodBuilder::new()
-                .metadata_builder(|m| {
-                    m.with_recommended_labels(build_recommended_labels(
-                        opa,
-                        &resolved_product_image.app_version_label,
-                        &rolegroup_ref.role,
-                        &rolegroup_ref.role_group,
-                    ))
-                })
-                .add_init_container(init_container)
-                .add_container(container_opa)
-                .add_container(container_bundle_builder)
-                .image_pull_secrets_from_product_image(resolved_product_image)
-                .add_volume(Volume {
-                    name: "config".to_string(),
-                    config_map: Some(ConfigMapVolumeSource {
-                        name: Some(rolegroup_ref.object_name()),
-                        ..ConfigMapVolumeSource::default()
-                    }),
-                    ..Volume::default()
-                })
-                .add_volume(Volume {
-                    name: "bundles".to_string(),
-                    empty_dir: Some(EmptyDirVolumeSource::default()),
-                    ..Volume::default()
-                })
-                .service_account_name(sa_name)
-                .security_context(PodSecurityContext {
-                    run_as_user: Some(1000),
-                    run_as_group: Some(1000),
-                    fs_group: Some(1000),
-                    ..PodSecurityContext::default()
-                })
-                .build_template(),
+            template: pb.build_template(),
             ..DaemonSetSpec::default()
         }),
         status: None,
@@ -592,19 +709,121 @@ bundles:
     persist: true
     polling:
       min_delay_seconds: 10
-      max_delay_seconds: 20"
+      max_delay_seconds: 20
+
+decision_logs:
+    console: true
+"
 }
 
-fn build_opa_start_command() -> Vec<String> {
-    vec![
-        "/stackable/opa/opa".to_string(),
-        "run".to_string(),
-        "-s".to_string(),
-        "-a".to_string(),
-        format!("0.0.0.0:{}", APP_PORT),
-        "-c".to_string(),
-        "/stackable/config/config.yaml".to_string(),
-    ]
+fn build_opa_start_command(merged_config: &OpaConfig, container_name: &str) -> String {
+    let mut opa_log_level = OpaLogLevel::Info;
+    let mut console_logging_off = false;
+
+    if let Some(ContainerLogConfig {
+        choice: Some(ContainerLogConfigChoice::Automatic(log_config)),
+    }) = merged_config.logging.containers.get(&Container::Opa)
+    {
+        // Retrieve the file log level for OPA and convert to OPA log levels
+        if let Some(AppenderConfig {
+            level: Some(log_level),
+        }) = log_config.file
+        {
+            opa_log_level = OpaLogLevel::from(log_level);
+        }
+
+        // We need to check if the console logging is deactivated (NONE)
+        // This will result in not using `tee` later on in the start command
+        if let Some(AppenderConfig {
+            level: Some(log_level),
+        }) = log_config.console
+        {
+            console_logging_off = log_level == LogLevel::NONE
+        }
+    }
+
+    let mut start_command = format!("/stackable/opa/opa run -s -a 0.0.0.0:{APP_PORT} -c {CONFIG_DIR}/config.yaml -l {opa_log_level}");
+
+    if console_logging_off {
+        start_command.push_str(&format!(" |& /stackable/multilog s{MAX_OPA_LOG_FILE_SIZE_IN_BYTES} n{OPA_ROLLING_LOG_FILES} {LOG_DIR}/{container_name}"));
+    } else {
+        start_command.push_str(&format!(" |& tee >(/stackable/multilog s{MAX_OPA_LOG_FILE_SIZE_IN_BYTES} n{OPA_ROLLING_LOG_FILES} {LOG_DIR}/{container_name})"));
+    }
+
+    start_command
+}
+
+fn build_bundle_builder_start_command(merged_config: &OpaConfig, container_name: &str) -> String {
+    let mut console_logging_off = false;
+
+    // We need to check if the console logging is deactivated (NONE)
+    // This will result in not using `tee` later on in the start command
+    if let Some(ContainerLogConfig {
+        choice: Some(ContainerLogConfigChoice::Automatic(log_config)),
+    }) = merged_config
+        .logging
+        .containers
+        .get(&Container::BundleBuilder)
+    {
+        if let Some(AppenderConfig {
+            level: Some(log_level),
+        }) = log_config.console
+        {
+            console_logging_off = log_level == LogLevel::NONE
+        }
+    };
+
+    let mut start_command = "/stackable/opa-bundle-builder".to_string();
+
+    if console_logging_off {
+        start_command.push_str(&format!(" |& /stackable/multilog s{MAX_OPA_BUNDLE_BUILDER_LOG_FILE_SIZE_IN_BYTES} n{OPA_ROLLING_BUNDLE_BUILDER_LOG_FILES} {LOG_DIR}/{container_name}"))
+    } else {
+        start_command.push_str(&format!(" |& tee >(/stackable/multilog s{MAX_OPA_BUNDLE_BUILDER_LOG_FILE_SIZE_IN_BYTES} n{OPA_ROLLING_BUNDLE_BUILDER_LOG_FILES} {LOG_DIR}/{container_name})"));
+    }
+
+    start_command
+}
+
+fn bundle_builder_log_level(merged_config: &OpaConfig) -> BundleBuilderLogLevel {
+    if let Some(ContainerLogConfig {
+        choice: Some(ContainerLogConfigChoice::Automatic(log_config)),
+    }) = merged_config
+        .logging
+        .containers
+        .get(&Container::BundleBuilder)
+    {
+        if let Some(logger) = log_config
+            .loggers
+            .get(AutomaticContainerLogConfig::ROOT_LOGGER)
+        {
+            return BundleBuilderLogLevel::from(logger.level);
+        }
+    }
+
+    BundleBuilderLogLevel::Info
+}
+
+fn build_prepare_start_command(merged_config: &OpaConfig, container_name: &str) -> Vec<String> {
+    let mut prepare_container_args = vec![];
+    if let Some(ContainerLogConfig {
+        choice: Some(ContainerLogConfigChoice::Automatic(log_config)),
+    }) = merged_config.logging.containers.get(&Container::Prepare)
+    {
+        prepare_container_args.push(product_logging::framework::capture_shell_output(
+            LOG_DIR,
+            container_name,
+            log_config,
+        ));
+    }
+
+    prepare_container_args.push(format!("echo \"Create dir [{BUNDLES_ACTIVE_DIR}]\""));
+    prepare_container_args.push(format!("mkdir -p {BUNDLES_ACTIVE_DIR}"));
+    prepare_container_args.push(format!("echo \"Create dir [{BUNDLES_INCOMING_DIR}]\""));
+    prepare_container_args.push(format!("mkdir -p {BUNDLES_INCOMING_DIR}"));
+    prepare_container_args.push(format!("echo \"Create dir [{BUNDLES_TMP_DIR}]\""));
+    prepare_container_args.push(format!("mkdir -p {BUNDLES_TMP_DIR}"));
+
+    prepare_container_args
 }
 
 fn service_ports() -> Vec<ServicePort> {